refactor: create shared utils package
