+++ packages/shared/pyproject.toml
@@ -1,0 +1,16 @@
+[tool.poetry]
+name = "shared-utils"
+version = "0.1.0"
+description = "Shared utility functions."
+authors = ["leomosley <leo.mosley2@gmail.com>"]
+license = "MIT"
+readme = "README.md"
+packages = [{ include = "shared_utils" }]
+
+[tool.poetry.dependencies]
+python = "^3.11"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
